refactor css-syntax-types into separate modules
